requirements.txt
boto3==1.35.19
botocore==1.35.19
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.3.7
gunicorn==23.0.0
h11==0.14.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
s3transfer==0.10.2
urllib3==2.2.3
Werkzeug==3.0.4